Move slurry recipes to BaseFluids
