+++ src/main/java/io/github/pylonmc/pylon/base/BaseFluids.java
@@ -4,6 +4,7 @@ import io.github.pylonmc.pylon.base.content.machines.smelting.Slurry;
 import io.github.pylonmc.pylon.base.recipes.CastingRecipe;
 import io.github.pylonmc.pylon.base.recipes.MeltingRecipe;
 import io.github.pylonmc.pylon.base.recipes.MixingPotRecipe;
+import io.github.pylonmc.pylon.base.recipes.SmelteryRecipe;
 import io.github.pylonmc.pylon.core.fluid.PylonFluid;
 import io.github.pylonmc.pylon.core.fluid.tags.FluidTemperature;
 import io.github.pylonmc.pylon.core.recipe.FluidOrItem;
@@ -12,6 +13,7 @@ import org.bukkit.NamespacedKey;
 import org.bukkit.inventory.ItemStack;
 
 import java.util.List;
+import java.util.Map;
 
 import static io.github.pylonmc.pylon.base.util.BaseUtils.baseKey;
 
@@ -90,16 +92,59 @@ public final class BaseFluids {
             baseKey("bronze"),
             Material.BROWN_CONCRETE
     ).addTag(FluidTemperature.HOT);
+    static {
+        SmelteryRecipe.RECIPE_TYPE.addRecipe(new SmelteryRecipe(
+                baseKey("bronze"),
+                Map.of(
+                        BaseFluids.COPPER, 1.0 - 0.12,
+                        BaseFluids.TIN, 0.12
+                ),
+                Map.of(BaseFluids.BRONZE, 1.0),
+                950
+        ));
+    }
 
     public static final PylonFluid BRASS = new PylonFluid(
             baseKey("brass"),
             Material.GOLD_BLOCK
     ).addTag(FluidTemperature.HOT);
+    static {
+        SmelteryRecipe.RECIPE_TYPE.addRecipe(new SmelteryRecipe(
+                baseKey("brass"),
+                Map.of(
+                        BaseFluids.COPPER, 1.0 - 0.3,
+                        BaseFluids.ZINC, 0.3
+                ),
+                Map.of(BaseFluids.BRASS, 1.0),
+                900
+        ));
+    }
 
     public static final PylonFluid STEEL = new PylonFluid(
             baseKey("steel"),
             Material.ORANGE_CONCRETE
     ).addTag(FluidTemperature.HOT);
+    static {
+        SmelteryRecipe.RECIPE_TYPE.addRecipe(new SmelteryRecipe(
+                baseKey("steel"),
+                Map.of(
+                        BaseFluids.IRON, 1.0 - 0.04,
+                        BaseFluids.CARBON_SLURRY, 0.04
+                ),
+                Map.of(BaseFluids.STEEL, 1.0),
+                1540
+        ));
+
+        SmelteryRecipe.RECIPE_TYPE.addRecipe(new SmelteryRecipe(
+                baseKey("decarburization"), // yes this is a real word
+                Map.of(
+                        BaseFluids.STEEL, 1.0,
+                        BaseFluids.WATER, 0.1
+                ),
+                Map.of(BaseFluids.IRON, 1.0),
+                1540
+        ));
+    }
 
     public static final PylonFluid SLURRY = new PylonFluid(
             baseKey("slurry"),
@@ -110,6 +155,17 @@ public final class BaseFluids {
             baseKey("slurry_coal"),
             BaseItems.COAL_DUST
     ).addTag(FluidTemperature.NORMAL);
+    static {
+        SmelteryRecipe.RECIPE_TYPE.addRecipe(new SmelteryRecipe(
+                baseKey("coal_to_carbon"),
+                Map.of(BaseFluids.COAL_SLURRY, 1.0),
+                Map.of(
+                        BaseFluids.CARBON_SLURRY, 0.9,
+                        BaseFluids.SLURRY, 0.1
+                ),
+                1000
+        ));
+    }
 
     public static final PylonFluid CARBON_SLURRY = new Slurry(
             baseKey("slurry_carbon"),
@@ -120,36 +176,165 @@ public final class BaseFluids {
             baseKey("slurry_raw_copper"),
             BaseItems.CRUSHED_RAW_COPPER
     ).addTag(FluidTemperature.NORMAL);
+    static {
+        SmelteryRecipe.RECIPE_TYPE.addRecipe(new SmelteryRecipe(
+                baseKey("copper_smelting"),
+                Map.of(BaseFluids.RAW_COPPER_SLURRY, 1.0),
+                Map.of(
+                        BaseFluids.COPPER, 0.5,
+                        BaseFluids.SLURRY, 0.5
+                ),
+                1085
+        ));
+
+        SmelteryRecipe.RECIPE_TYPE.addRecipe(new SmelteryRecipe(
+                baseKey("copper_smelting_with_sulfur"),
+                Map.of(
+                        BaseFluids.RAW_COPPER_SLURRY, 1.0,
+                        BaseFluids.SULFUR, 0.1
+                ),
+                Map.of(
+                        BaseFluids.COPPER, 0.5,
+                        BaseFluids.SLURRY, 0.3,
+                        BaseFluids.RAW_LEAD_SLURRY, 0.1,
+                        BaseFluids.RAW_ZINC_SLURRY, 0.1
+                ),
+                1085
+        ));
+    }
 
     public static final PylonFluid RAW_GOLD_SLURRY = new Slurry(
             baseKey("slurry_raw_gold"),
             BaseItems.CRUSHED_RAW_GOLD
     ).addTag(FluidTemperature.NORMAL);
+    static {
+        SmelteryRecipe.RECIPE_TYPE.addRecipe(new SmelteryRecipe(
+                baseKey("gold_smelting"),
+                Map.of(
+                        BaseFluids.RAW_GOLD_SLURRY, 1.0,
+                        BaseFluids.MERCURY, 1.0
+                ),
+                Map.of(
+                        BaseFluids.GOLD, 0.5,
+                        BaseFluids.SLURRY, 0.3,
+                        BaseFluids.SILVER, 0.15,
+                        BaseFluids.RAW_TIN_SLURRY, 0.15,
+                        BaseFluids.MERCURY, 0.9
+                ),
+                1064
+        ));
+    }
 
     public static final PylonFluid RAW_IRON_SLURRY = new Slurry(
             baseKey("slurry_raw_iron"),
             BaseItems.CRUSHED_RAW_IRON
     ).addTag(FluidTemperature.NORMAL);
+    static {
+        SmelteryRecipe.RECIPE_TYPE.addRecipe(new SmelteryRecipe(
+                baseKey("iron_smelting"),
+                Map.of(
+                        BaseFluids.RAW_IRON_SLURRY, 1.0,
+                        BaseFluids.CARBON_SLURRY, 0.5
+                ),
+                Map.of(
+                        BaseFluids.IRON, 1.0,
+                        BaseFluids.SLURRY, 0.5
+                ),
+                1540
+        ));
+
+        SmelteryRecipe.RECIPE_TYPE.addRecipe(new SmelteryRecipe(
+                baseKey("iron_smelting_with_sulfur"),
+                Map.of(
+                        BaseFluids.RAW_IRON_SLURRY, 1.0,
+                        BaseFluids.CARBON_SLURRY, 0.5,
+                        BaseFluids.SULFUR, 0.1
+                ),
+                Map.of(
+                        BaseFluids.IRON, 1.0,
+                        BaseFluids.SLURRY, 0.4,
+                        BaseFluids.COBALT, 0.1,
+                        BaseFluids.NICKEL, 0.1
+                ),
+                1540
+        ));
+    }
 
     public static final PylonFluid RAW_LEAD_SLURRY = new Slurry(
             baseKey("slurry_raw_lead"),
             BaseItems.CRUSHED_RAW_LEAD
     ).addTag(FluidTemperature.NORMAL);
+    static {
+        SmelteryRecipe.RECIPE_TYPE.addRecipe(new SmelteryRecipe(
+                baseKey("lead_smelting"),
+                Map.of(
+                        BaseFluids.RAW_LEAD_SLURRY, 1.0,
+                        BaseFluids.CARBON_SLURRY, 0.5,
+                        BaseFluids.SULFUR, 0.1
+                ),
+                Map.of(
+                        BaseFluids.LEAD, 1.0,
+                        BaseFluids.SLURRY, 0.5,
+                        BaseFluids.SILVER, 0.1
+                ),
+                350
+        ));
+    }
 
     public static final PylonFluid RAW_TIN_SLURRY = new Slurry(
             baseKey("slurry_raw_tin"),
             BaseItems.CRUSHED_RAW_TIN
     ).addTag(FluidTemperature.NORMAL);
+    static {
+        SmelteryRecipe.RECIPE_TYPE.addRecipe(new SmelteryRecipe(
+                baseKey("tin_smelting"),
+                Map.of(
+                        BaseFluids.RAW_TIN_SLURRY, 1.0,
+                        BaseFluids.CARBON_SLURRY, 0.5
+                ),
+                Map.of(
+                        BaseFluids.TIN, 1.0,
+                        BaseFluids.SLURRY, 0.5
+                ),
+                250
+        ));
+    }
 
     public static final PylonFluid RAW_ZINC_SLURRY = new Slurry(
             baseKey("slurry_raw_zinc"),
             BaseItems.CRUSHED_RAW_ZINC
     ).addTag(FluidTemperature.NORMAL);
+    static {
+        SmelteryRecipe.RECIPE_TYPE.addRecipe(new SmelteryRecipe(
+                baseKey("zinc_smelting"),
+                Map.of(
+                        BaseFluids.RAW_ZINC_SLURRY, 1.0,
+                        BaseFluids.CARBON_SLURRY, 0.5
+                ),
+                Map.of(
+                        BaseFluids.ZINC, 1.0,
+                        BaseFluids.SLURRY, 0.5
+                ),
+                700
+        ));
+    }
 
     public static final PylonFluid REDSTONE_SLURRY = new Slurry(
             baseKey("slurry_redstone"),
             new ItemStack(Material.REDSTONE)
     ).addTag(FluidTemperature.NORMAL);
+    static {
+        SmelteryRecipe.RECIPE_TYPE.addRecipe(new SmelteryRecipe(
+                baseKey("redstone_decomposition"),
+                Map.of(BaseFluids.REDSTONE_SLURRY, 1.0),
+                Map.of(
+                        BaseFluids.SULFUR, 0.25,
+                        BaseFluids.MERCURY, 0.25,
+                        BaseFluids.SLURRY, 0.5
+                ),
+                345
+        ));
+    }
 
     public static final PylonFluid PLANT_OIL = new PylonFluid(
             baseKey("plant_oil"),
